requirements.txt
click
rich
python-nmap
requests
